Split authentication into a helper dataclass.
Deprecated the old-school method of passing pkcs12s_filename/pass
directly. Instead, you now pass an `auth_type` dataclass that
encodes this information. Current auth types are CertificateAuth
or UsernamePassAuth.

For CertificateAuth, you can pass a direct byte array for the
certificate, instead of a filename if desired.

Updated the examples in the README and wrote a tentative CHANGELOG
entry.